requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
coverage==7.3.2
discord.py==2.3.2
exceptiongroup==1.2.0
frozenlist==1.4.0
idna==3.6
iniconfig==2.0.0
multidict==6.0.4
mysql-connector-python==8.2.0
packaging==23.2
pluggy==1.3.0
protobuf==4.21.12
pytest==7.4.3
pytest-atomic==2.0.0
pytest-cov==4.1.0
python-dotenv==1.0.0
tomli==2.0.1
yarl==1.9.3
